+++ src/schemas/sdo/log-source.schema.ts
@@ -8,20 +8,49 @@ import {
   xMitreModifiedByRefSchema,
   xMitrePlatformsSchema,
 } from '../common/common-properties.js';
-import { stixCreatedByRefSchema } from '../common/misc.js';
+import { externalReferenceSchema, stixCreatedByRefSchema } from '../common/misc.js';
 import { createStixIdentifierSchema } from '../common/stix-identifier.js';
 import { stixTypeSchema } from '../common/stix-type.js';
-import { xMitreCollectionLayersSchema } from './data-source.schema.js';
+import { MitreCollectionLayersOV } from '../common/open-vocabulary.js';
 
-export const logSource = attackBaseObjectSchema.extend({
+
+/////////////////////////////////////
+//
+// MITRE Collection Layers
+// (x_mitre_collection_layers)
+//
+/////////////////////////////////////
+
+export const xMitreCollectionLayersSchema = z
+.array(MitreCollectionLayersOV, {
+  invalid_type_error:
+  'x_mitre_collection_layers must be an array of supported collection layers.',
+})
+.describe('List of places the data can be collected from.');
+
+export type XMitreCollectionLayers = z.infer<typeof xMitreCollectionLayersSchema>;
+
+
+/////////////////////////////////////
+//
+// MITRE Log Source SDO
+// (x-mitre-log-source)
+//
+/////////////////////////////////////
+
+export const logSourceSchema = attackBaseObjectSchema.extend({
   id: createStixIdentifierSchema('x-mitre-log-source'),
 
   type: z.literal(stixTypeSchema.enum['x-mitre-log-source']),
 
+  // Optional in STIX but required in ATT&CK
   created_by_ref: stixCreatedByRefSchema,
 
   description: descriptionSchema,
 
+  external_references: externalReferenceSchema,
+
+  // Optional in STIX but required in ATT&CK
   object_marking_refs: objectMarkingRefsSchema,
 
   x_mitre_platforms: xMitrePlatformsSchema.optional(),
@@ -33,4 +62,31 @@ export const logSource = attackBaseObjectSchema.extend({
   x_mitre_contributors: xMitreContributorsSchema.optional(),
 
   x_mitre_collection_layers: xMitreCollectionLayersSchema,
-});
+})
+  .strict()
+  .superRefine((schema, ctx) => {
+    //==============================================================================
+    // Validate external references
+    //==============================================================================
+
+    const { external_references } = schema;
+    const attackIdEntry = external_references[0];
+    if (!attackIdEntry.external_id) {
+      ctx.addIssue({
+        code: z.ZodIssueCode.custom,
+        message: 'ATT&CK ID must be defined in the first external_references entry.',
+        path: ['external_references', 0, 'external_id'],
+      });
+    } else {
+      const idRegex = /^DS\d{4}$/;
+      if (!idRegex.test(attackIdEntry.external_id)) {
+        ctx.addIssue({
+          code: z.ZodIssueCode.custom,
+          message: `The first external_reference must match the ATT&CK ID format DS####.`,
+          path: ['external_references', 0, 'external_id'],
+        });
+      }
+    }
+  });
+
+export type LogSource = z.infer<typeof logSourceSchema>;
