+++ src/schemas/sdo/stix-bundle.schema.ts
@@ -6,7 +6,7 @@ import { type Malware, malwareSchema } from './malware.schema.js';
 import { type Asset, assetSchema } from './asset.schema.js';
 import { type Campaign, campaignSchema } from './campaign.schema.js';
 import { type DataComponent, dataComponentSchema } from './data-component.schema.js';
-import { type DataSource, dataSourceSchema } from './data-source.schema.js';
+import { type LogSource, logSourceSchema } from './log-source.schema.js';
 import { type Identity, identitySchema } from './identity.schema.js';
 import { type Matrix, matrixSchema } from './matrix.schema.js';
 import { type Tool, toolSchema } from './tool.schema.js';
@@ -31,7 +31,7 @@ export type AttackObject =
   | Campaign
   | Collection
   | DataComponent
-  | DataSource
+  | LogSource
   | Identity
   | Matrix
   | Tool
@@ -53,7 +53,7 @@ export const attackObjectsSchema: z.ZodTypeAny = z
       campaignSchema,
       collectionSchema,
       dataComponentSchema,
-      dataSourceSchema,
+      logSourceSchema,
       identitySchema,
       matrixSchema,
       toolSchema,
